Work around items with empty requests or responses.
I had a fairly large (several GB, hundreds of thousands of records)
history created with Burp 1.7.23.  Sometimes the server had crashed or
otherwise failed to ever respond to Burp, and the saved response was
empty:

    <responselength></responselength>
    <mimetype></mimetype>
    <response base64="true"></response>

convert-burp-suite-http-proxy-history-to-csv.py was blowing up the
first time it encountered such a case, with an error that boiled
down to:

  line 70, in convert_to_output_file
      format_handler.row_column(line['response']['#text'], encoded=True)
  KeyError: u'#text'

I am a useless Python programmer but it seemed to me that when the
response was empty, the key #text was undefined, causing this error.
This patch adds sanity checking to work around these; I check
'request' too although I do not know that that can ever be blank.